Update shim to include GUID
Bootkit now includes the uki guids when computing the vendor_db.esl
for shim. Thus we need to include the uki guids when computing the
pcr7 values for a keyset. This commit does that.

This commit also adds ability for trust computePCR7 to output
the hash calculations of predected tpm events for pcr7. The output
can go to stdout or to a file.

Signed-off-by: Joy Latten <jlatten@cisco.com>